refactor: 테스트용 유틸 패키지 이동
