+++ src/test/java/domain/ResultAnalyzerTest.java
@@ -9,6 +9,7 @@ import domain.player.Players;
 import domain.analyzer.dto.ResultAnalysisDto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import utils.CardDeckBuilder;
 
 import java.util.List;
