+++ src/test/java/domain/card/CardBundleTest.java
@@ -3,6 +3,7 @@ package domain.card;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import utils.CardBundleBuilder;
 
 public class CardBundleTest {
