+++ src/test/java/domain/card/CardDeckTest.java
@@ -3,6 +3,7 @@ package domain.card;
 import domain.card.exception.CardException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import utils.CardDeckBuilder;
 
 public class CardDeckTest {
