+++ src/test/java/domain/dealer/DealerTest.java
@@ -6,6 +6,7 @@ import domain.player.Player;
 import domain.player.PlayerName;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import utils.CardDeckBuilder;
 
 import java.util.List;
