+++ src/test/java/domain/player/PlayerTest.java
@@ -6,6 +6,7 @@ import domain.dealer.Dealer;
 import domain.player.exception.PlayerException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import utils.CardDeckBuilder;
 
 import java.util.List;
